facial-expression-api/requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
numpy
pillow
pandas
scikit-learn
matplotlib
kagglehub
